Redundancy fixes
Added while loop to reduce key if it is over 26. Since the alphabet is 26 letters long, there is no need to have the key be over 26. Also fixed assignment of current letter as the new values in cipherTxt/decipherTxt. No need to assign within if-statements that are just for assigning values to currentLetter